refactor: separate NavigationBar into modular desktop/mobile components with type safety
