+++ web/src/layout/public/MainBody.tsx
@@ -5,5 +5,5 @@ interface MainBodyProps {
 }
 
 export const MainBody: React.FC<MainBodyProps> = ({ children }) => {
-      return <main className='max-w-12xl mx-auto  bg-cream   '>{children}</main>;
+      return <main className='max-w-12xl mx-auto mt-16 sm:mt-20  bg-cream'>{children}</main>;
 };
